+++ AppUtility.h
@@ -19,8 +19,8 @@ void Logging(QtMsgType type, const QMessageLogContext& context, const QString& m
 
 void OpenLocalPath(const QString&);
 
-double ToMByte(int sizeInBit);
-double ToKByte(int sizeInBit);
+double ToMByte(long long sizeInBit);
+double ToKByte(long long sizeInBit);
 
 void ShowWarningBox(const QString& title, const QString& waring, const QString& btnText);
 bool ShowQuestionBox(const QString& title, const QString& info, const QString& yesText, const QString& noText);
